make SPDK installation step optional
Omit installation of SPDK to support projects that do not use it.
This brings more versatility how this Github Action can be applied.
SPDK installation can be skipped by either:
- not setting SPDK version variable
- setting SPDK version variable to an empty string
- setting SPDK version variable to "none"

+++ main.sh
@@ -3,12 +3,8 @@ set -e
 set -o pipefail
 
 CODEROOT=$HOME/setup-dpdk
-mkdir -p $CODEROOT/dpdk_$DPDKVER $CODEROOT/spdk_$SPDKVER
+mkdir -p $CODEROOT/dpdk_$DPDKVER
 
-cd $CODEROOT/spdk_$SPDKVER
-if ! [[ -f scripts/pkgdep.sh ]]; then
-  curl -sfL https://github.com/spdk/spdk/archive/v$SPDKVER.tar.gz | tar -xz --strip-components=1
-fi
 sudo apt-get install python3-pyelftools python3-setuptools
 sudo scripts/pkgdep.sh
 
@@ -29,15 +25,23 @@ sudo ninja -C build install
 sudo find /usr/local/lib -name 'librte_*.a' -delete
 sudo ldconfig
 
-cd $CODEROOT/spdk_$SPDKVER
-if ! [[ -f build/lib/libspdk_env_dpdk.a ]]; then
-  ./configure --target-arch=$TARGETARCH --enable-debug --disable-tests --with-shared \
-    --with-dpdk=/usr/local --without-vhost --without-isal --without-fuse
-  make -j$(nproc)
+if [[ ! -z $SPDKVER && -n $SPDKVER && $SPDKVER != "none" ]]; then
+  mkdir -p  $CODEROOT/spdk_$SPDKVER
+  cd $CODEROOT/spdk_$SPDKVER
+  if ! [[ -f scripts/pkgdep.sh ]]; then
+    curl -sfL https://github.com/spdk/spdk/archive/v$SPDKVER.tar.gz | tar -xz --strip-components=1
+  fi
+
+  cd $CODEROOT/spdk_$SPDKVER
+  if ! [[ -f build/lib/libspdk_env_dpdk.a ]]; then
+    ./configure --target-arch=$TARGETARCH --enable-debug --disable-tests --with-shared \
+      --with-dpdk=/usr/local --without-vhost --without-isal --without-fuse
+    make -j$(nproc)
+  fi
+  sudo make install
+  sudo find /usr/local/lib -name 'libspdk_*.a' -delete
+  sudo ldconfig
 fi
-sudo make install
-sudo find /usr/local/lib -name 'libspdk_*.a' -delete
-sudo ldconfig
 
 if [[ $NRHUGE -gt 0 ]]; then
   echo $NRHUGE | sudo tee /sys/devices/system/node/node*/hugepages/hugepages-2048kB/nr_hugepages >/dev/null
